Rework our netstandard1.3 test references
The `netstandard1.3` test references were being provided by our
MS.CA.Test.Resources.Proprietary package. The contents of this package
get included in the output directory of every single unit test DLL that
we build (which contributes to build output size). The references
themselves though were only used in a very small number of tests.

This change upgrades to a version of MS.CA.Test.Resources.Proprietary
that has the `netstandard1.3` references removed and adds the
`netstandard1.3` references to only the projects that needs them
