+++ src/Compilers/Test/Core/Mocks/TestReferences.cs
@@ -104,14 +104,6 @@ public static class TestReferences
         }
     }
 
-    public static class NetStandard13
-    {
-        private static readonly Lazy<PortableExecutableReference> s_systemRuntime = new Lazy<PortableExecutableReference>(
-            () => AssemblyMetadata.CreateFromImage(ProprietaryTestResources.netstandard13.System_Runtime).GetReference(display: @"System.Runtime.dll (netstandard13 ref)"),
-            LazyThreadSafetyMode.PublicationOnly);
-        public static PortableExecutableReference SystemRuntime => s_systemRuntime.Value;
-    }
-
     public static class DiagnosticTests
     {
         public static class ErrTestLib01
