+++ src/Compilers/Test/Core/Platform/Desktop/TestHelpers.cs
@@ -93,196 +93,6 @@ public class TestAnalyzer : DiagnosticAnalyzer
             return dir.CreateFile(assemblyName + ".dll").WriteAllBytes(analyzerCompilation.EmitToArray());
         }
 
-        public static ImmutableArray<byte> CreateCSharpAnalyzerNetStandard13(string analyzerAssemblyName)
-        {
-            var minSystemCollectionsImmutableSource = @"
-[assembly: System.Reflection.AssemblyVersion(""1.2.3.0"")]
-
-namespace System.Collections.Immutable
-{
-    public struct ImmutableArray<T>
-    {
-    }
-}
-";
-
-            var minCodeAnalysisSource = @"
-using System;
-
-[assembly: System.Reflection.AssemblyVersion(""2.0.0.0"")]
-
-namespace Microsoft.CodeAnalysis.Diagnostics
-{
-    [AttributeUsage(AttributeTargets.Class)]
-    public sealed class DiagnosticAnalyzerAttribute : Attribute
-    {
-        public DiagnosticAnalyzerAttribute(string firstLanguage, params string[] additionalLanguages) {}
-    }
-
-    public abstract class DiagnosticAnalyzer
-    {
-        public abstract System.Collections.Immutable.ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics { get; }
-        public abstract void Initialize(AnalysisContext context);
-    }
-
-    public abstract class AnalysisContext
-    {
-    }
-}
-
-namespace Microsoft.CodeAnalysis
-{
-    public sealed class DiagnosticDescriptor
-    {
-    }
-}
-";
-            var minSystemCollectionsImmutableImage = CSharpCompilation.Create(
-                "System.Collections.Immutable",
-                new[] { SyntaxFactory.ParseSyntaxTree(minSystemCollectionsImmutableSource) },
-                new[] { MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Runtime) },
-                new CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary, cryptoPublicKey: TestResources.TestKeys.PublicKey_b03f5f7f11d50a3a)).EmitToArray();
-
-            var minSystemCollectionsImmutableRef = MetadataReference.CreateFromImage(minSystemCollectionsImmutableImage);
-
-            var minCodeAnalysisImage = CSharpCompilation.Create(
-                "Microsoft.CodeAnalysis",
-                new[] { SyntaxFactory.ParseSyntaxTree(minCodeAnalysisSource) },
-                new[]
-                {
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Runtime),
-                    minSystemCollectionsImmutableRef
-                },
-                new CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary, cryptoPublicKey: TestResources.TestKeys.PublicKey_31bf3856ad364e35)).EmitToArray();
-
-            var minCodeAnalysisRef = MetadataReference.CreateFromImage(minCodeAnalysisImage);
-
-            var analyzerSource = @"
-using System;
-using System.Collections.ObjectModel;
-using System.Collections.Immutable;
-using System.ComponentModel;
-using System.Diagnostics;
-using System.Globalization;
-using System.IO;
-using System.IO.Compression;
-using System.Net.Http;
-using System.Net.Security;
-using System.Net.Sockets;
-using System.Reflection;
-using System.Runtime.InteropServices;
-using System.Runtime.Serialization;
-using System.Security.AccessControl;
-using System.Security.Cryptography;
-using System.Security.Principal;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
-using System.Xml;
-using System.Xml.Serialization;
-using System.Xml.XPath;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.Win32.SafeHandles;
-
-[DiagnosticAnalyzer(""C#"")]
-public class TestAnalyzer : DiagnosticAnalyzer
-{
-    public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => throw new NotImplementedException(new[]                                 
-    {                                                                                                                                   
-        typeof(Win32Exception),           // Microsoft.Win32.Primitives
-        typeof(AppContext),               // System.AppContext
-        typeof(Console),                  // System.Console
-        typeof(ValueTuple),               // System.ValueTuple
-        typeof(FileVersionInfo),          // System.Diagnostics.FileVersionInfo
-        typeof(Process),                  // System.Diagnostics.Process
-        typeof(ChineseLunisolarCalendar), // System.Globalization.Calendars
-        typeof(ZipArchive),               // System.IO.Compression
-        typeof(ZipFile),                  // System.IO.Compression.ZipFile
-        typeof(FileOptions),              // System.IO.FileSystem
-        typeof(FileAttributes),           // System.IO.FileSystem.Primitives
-        typeof(HttpClient),               // System.Net.Http
-        typeof(AuthenticatedStream),      // System.Net.Security
-        typeof(IOControlCode),            // System.Net.Sockets
-        typeof(RuntimeInformation),       // System.Runtime.InteropServices.RuntimeInformation
-        typeof(SerializationException),   // System.Runtime.Serialization.Primitives
-        typeof(GenericIdentity),          // System.Security.Claims
-        typeof(Aes),                      // System.Security.Cryptography.Algorithms
-        typeof(CspParameters),            // System.Security.Cryptography.Csp
-        typeof(AsnEncodedData),           // System.Security.Cryptography.Encoding
-        typeof(AsymmetricAlgorithm),      // System.Security.Cryptography.Primitives
-        typeof(SafeX509ChainHandle),      // System.Security.Cryptography.X509Certificates
-        typeof(IXmlLineInfo),             // System.Xml.ReaderWriter
-        typeof(XmlNode),                  // System.Xml.XmlDocument
-        typeof(XPathDocument),            // System.Xml.XPath
-        typeof(XDocumentExtensions),      // System.Xml.XPath.XDocument
-        typeof(CodePagesEncodingProvider),// System.Text.Encoding.CodePages
-        typeof(ValueTask<>),              // System.Threading.Tasks.Extensions
-
-        // csc doesn't ship with facades for the following assemblies. 
-        // Analyzers can't use them unless they carry the facade with them.
-
-        // typeof(SafePipeHandle),           // System.IO.Pipes
-        // typeof(StackFrame),               // System.Diagnostics.StackTrace
-        // typeof(BindingFlags),             // System.Reflection.TypeExtensions
-        // typeof(AccessControlActions),     // System.Security.AccessControl
-        // typeof(SafeAccessTokenHandle),    // System.Security.Principal.Windows
-        // typeof(Thread),                   // System.Threading.Thread
-    }.Length.ToString());
-
-    public override void Initialize(AnalysisContext context)
-    {
-    }
-}";
-
-            var analyzerImage = CSharpCompilation.Create(
-                analyzerAssemblyName,
-                new SyntaxTree[] { SyntaxFactory.ParseSyntaxTree(analyzerSource) },
-                new MetadataReference[]
-                {
-                    minCodeAnalysisRef,
-                    minSystemCollectionsImmutableRef,
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.Microsoft_Win32_Primitives),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_AppContext),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Console),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_ValueTuple),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Diagnostics_FileVersionInfo),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Diagnostics_Process),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Diagnostics_StackTrace),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Globalization_Calendars),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_IO_Compression),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_IO_Compression_ZipFile),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_IO_FileSystem),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_IO_FileSystem_Primitives),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_IO_Pipes),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Net_Http),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Net_Security),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Net_Sockets),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Reflection_TypeExtensions),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Runtime),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Runtime_InteropServices_RuntimeInformation),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Runtime_Serialization_Primitives),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Security_AccessControl),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Security_Claims),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Security_Cryptography_Algorithms),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Security_Cryptography_Csp),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Security_Cryptography_Encoding),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Security_Cryptography_Primitives),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Security_Cryptography_X509Certificates),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Security_Principal_Windows),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Threading_Thread),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Threading_Tasks_Extensions),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Xml_ReaderWriter),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Xml_XmlDocument),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Xml_XPath),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Xml_XPath_XDocument),
-                    MetadataReference.CreateFromImage(ProprietaryTestResources.netstandard13.System_Text_Encoding_CodePages)
-                },
-                new CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary)).EmitToArray();
-
-            return analyzerImage;
-        }
-
         public static string? GetMSBuildDirectory()
         {
             var vsVersion = Environment.GetEnvironmentVariable("VisualStudioVersion") ?? "14.0";
