+++ src/Compilers/Test/Core/TestBase.cs
@@ -250,11 +250,6 @@ namespace Roslyn.Test.Utilities
 
         public static MetadataReference Net46StandardFacade => s_46NetStandardFacade.Value;
 
-        private static readonly Lazy<MetadataReference> s_systemDynamicRuntimeRef = new Lazy<MetadataReference>(
-            () => AssemblyMetadata.CreateFromImage(ProprietaryTestResources.netstandard13.System_Dynamic_Runtime).GetReference(display: "System.Dynamic.Runtime.dll (netstandard 1.3 ref)"),
-            LazyThreadSafetyMode.PublicationOnly);
-        public static MetadataReference SystemDynamicRuntimeRef => s_systemDynamicRuntimeRef.Value;
-
         private static readonly Lazy<MetadataReference> s_systemRef = new Lazy<MetadataReference>(
             () => AssemblyMetadata.CreateFromImage(ResourcesNet451.System).GetReference(display: "System.v4_0_30319.dll"),
             LazyThreadSafetyMode.PublicationOnly);
